fix(accordion)!: remove BaseAccordion
Closes #2612

fix(accordion): expand logic

refactor(accordion): prevent circular imports
