+++ elements/pf-accordion/pf-accordion-panel.ts
@@ -1,13 +1,13 @@
+import { LitElement, html } from 'lit';
 import { customElement } from 'lit/decorators/custom-element.js';
 import { property } from 'lit/decorators/property.js';
 
-import { BaseAccordionPanel } from './BaseAccordionPanel.js';
+import { getRandomId } from '@patternfly/pfe-core/functions/random.js';
 
 import style from './pf-accordion-panel.css';
 
 /**
  * Accordion Panel
- *
  * @slot - Panel content
  * @cssprop     {<color>} --pf-c-accordion--BackgroundColor
  *              Sets the background color for the panel content.
@@ -45,10 +45,30 @@ import style from './pf-accordion-panel.css';
  *              {@default `var(--pf-global--BorderWidth--lg, 3px)`}
  */
 @customElement('pf-accordion-panel')
-export class PfAccordionPanel extends BaseAccordionPanel {
-  static readonly styles = [...BaseAccordionPanel.styles, style];
+export class PfAccordionPanel extends LitElement {
+  static readonly styles = [style];
+
+  @property({ type: Boolean, reflect: true }) expanded = false;
 
   @property({ reflect: true }) bordered?: 'true' | 'false';
+
+  override connectedCallback() {
+    super.connectedCallback();
+    this.id ||= getRandomId(this.localName);
+    this.setAttribute('role', 'region');
+  }
+
+  override render() {
+    return html`
+      <div tabindex="-1">
+        <div id="container" class="content" part="container">
+          <div class="body">
+            <slot></slot>
+          </div>
+        </div>
+      </div>
+    `;
+  }
 }
 
 declare global {
